Hide top bar when UNSUSPEND_ON_FOCUS is enabled
The preview was not properly aligned with the page because of the top bar.
It gave the impression of a visual glitch when the tab was automatically
unsuspended.

This is fixed by collapsing the gsTopBar element (display=none) and
by clearing the top margin of the gsPreviewImg element.

When UNSUSPEND_ON_FOCUS is disabled, the top bar has to stay to give
a visual feedback to the user.