Add support for a "close all suspended tabs" option
This option allows the user to close all suspended tabs.

As a user, I often end up with many suspended tabs with a few
unsuspended tabs mixed in.  This option has saved me a bunch of
headache having to search for all of the active tabs when I want
to clean up the old unused tabs.