Support for HTML5 form inputs, adds checker to see if inputs still contain unsaved values
Expands list to include newer HTML5 input methods like what Facebook uses for comments:

```HTML
<div aria-autocomplete="list" aria-multiline="true" class="notranslate _5rpu" contenteditable="true" role="textbox" spellcheck="true" style="outline: none; user-select: text; white-space: pre-wrap; word-wrap: break-word;">
    Typed text appears here.
</div>
```

Also noticing that there doesn't appear to be anything in place to detect if the form/input was ever submitted. So I went ahead and added something to check if the elements that were edited haven't been removed and still have values set to the controls. 

UNTESTED